fix: remove common helper
